Move rcb::work_share into its own module
It can be useful for other algorithms, and is not really specific to RCB
